Slider change value on scroll only when mouse hover
This prevent from scrolling with the mouse and changing slider when the mouse is not hovering the slider.
For example: If we have multiple slider in a tab if we changed value of a slider by clicking with the mouse and the hover another slider an scroll, the two would change values.